refactor JniGenTask to Config
