+++ jni_gen/bin/jni_gen.dart
@@ -2,6 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:jni_gen/src/config/config.dart';
+import 'package:jni_gen/jni_gen.dart';
 
-void main(List<String> args) async => await JniGenTask.parseArgs(args).run();
+void main(List<String> args) async {
+  final config = Config.parseArgs(args);
+  await generateJniBindings(config);
+}
