+++ jni_gen/lib/src/summary/summary.dart
@@ -5,10 +5,6 @@
 import 'dart:io';
 import 'package:jni_gen/src/util/command_output.dart';
 
-abstract class SummarySource {
-  Future<Stream<List<int>>> getInputStream();
-}
-
 /// A command based summary source which calls the ApiSummarizer command.
 /// [sourcePaths] and [classPaths] can be provided for the summarizer to find
 /// required dependencies. The [classes] argument specifies the fully qualified
@@ -21,17 +17,17 @@ abstract class SummarySource {
 ///
 /// The default summarizer needs to be built with `jni_gen:setup`
 /// script before this API is used.
-class SummarizerCommand extends SummarySource {
+class SummarizerCommand {
   SummarizerCommand({
     this.command = "java -jar .dart_tool/jni_gen/ApiSummarizer.jar",
-    List<Uri>? sourcePaths,
-    List<Uri>? classPaths,
+    List<Uri>? sourcePath,
+    List<Uri>? classPath,
     this.extraArgs = const [],
     required this.classes,
     this.workingDirectory,
     this.backend,
-  })  : sourcePaths = sourcePaths ?? [],
-        classPaths = classPaths ?? [] {
+  })  : sourcePaths = sourcePath ?? [],
+        classPaths = classPath ?? [] {
     if (backend != null && !{'asm', 'doclet'}.contains(backend)) {
       throw ArgumentError('Supported backends: asm, doclet');
     }
@@ -70,7 +66,6 @@ class SummarizerCommand extends SummarySource {
     }
   }
 
-  @override
   Future<Stream<List<int>>> getInputStream() async {
     final commandSplit = command.split(" ");
     final exec = commandSplit[0];
