+++ jni_gen/test/test_util/test_util.dart
@@ -1,13 +1,9 @@
 import 'dart:io';
 
 import 'package:path/path.dart' hide equals;
-import 'package:jni_gen/jni_gen.dart';
-import 'package:jni_gen/tools.dart';
 import 'package:test/test.dart';
 
-const packageTestsDir = 'test';
-
-Future<bool> isEmptyDir(String path) async {
+Future<bool> hasNoFilesInDir(String path) async {
   final dir = Directory(path);
   return (!await dir.exists()) || (await dir.list().length == 0);
 }
@@ -22,12 +18,6 @@ Future<int> runCmd(String exec, List<String> args,
   return proc.exitCode;
 }
 
-Future<void> buildNativeLibs(String testName) async {
-  final testRoot = join(packageTestsDir, testName);
-  await runCmd('dart', ['run', 'jni:setup']);
-  await runCmd('dart', ['run', 'jni:setup', '-S', join(testRoot, 'src')]);
-}
-
 Future<List<String>> getJarPaths(String testRoot) {
   final jarPath = join(testRoot, 'jar');
   return Directory(jarPath)
@@ -35,55 +25,6 @@ Future<List<String>> getJarPaths(String testRoot) {
       .map((entry) => entry.path)
       .where((path) => path.endsWith('jar'))
       .toList();
-}
-
-/// Download dependencies using maven and generate bindings.
-Future<void> generateBindings({
-  required String testName,
-  required List<String> sourceDepNames,
-  required List<String> jarDepNames,
-  required List<String> classes,
-  required WrapperOptions options,
-  required bool isGeneratedFileTest,
-  bool useAsmBackend = false,
-  bool isThirdParty = false,
-  String? preamble,
-}) async {
-  final testRoot =
-      join(packageTestsDir, testName, isThirdParty ? 'third_party' : '');
-  final jarPath = join(testRoot, 'jar');
-  final javaPath = join(testRoot, 'java');
-  final src = join(testRoot, isGeneratedFileTest ? 'test_src' : 'src');
-  final lib = join(testRoot, isGeneratedFileTest ? 'test_lib' : 'lib');
-
-  final sourceDeps = MavenTools.deps(sourceDepNames);
-  final jarDeps = MavenTools.deps(jarDepNames);
-
-  MavenTools.setVerbose(true);
-  if (await isEmptyDir(jarPath)) {
-    await Directory(jarPath).create(recursive: true);
-    await MavenTools.downloadMavenJars(jarDeps, jarPath);
-  }
-  if (await isEmptyDir(javaPath)) {
-    await Directory(javaPath).create(recursive: true);
-    await MavenTools.downloadMavenSources(sourceDeps, javaPath);
-  }
-  final jars = await getJarPaths(testRoot);
-  stderr.writeln('using classpath: $jars');
-  await JniGenTask.ofComponents(
-          summarizer: SummarizerCommand(
-            sourcePaths: [Uri.directory(javaPath)],
-            classPaths: jars.map(Uri.file).toList(),
-            classes: classes,
-            extraArgs: useAsmBackend ? ['--backend', 'asm'] : [],
-          ),
-          options: options,
-          outputWriter: FilesWriter(
-              cWrapperDir: Uri.directory(src),
-              dartWrappersRoot: Uri.directory(lib),
-              preamble: preamble,
-              libraryName: testName))
-      .run();
 }
 
 /// compares 2 hierarchies, with and without prefix 'test_'
